Move ConnectionViewModel to viewmodel package
This commit moves the `ConnectionViewModel` class from the `com.maary.liveinpeace` package to the `com.maary.liveinpeace.viewmodel` package.

The import statements in `HistoryActivity.kt` have been updated accordingly.
